requirements.txt
# Telegram Bot Framework
python-telegram-bot==20.7

# Google APIs
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# HTTP and Web Framework (for webhook and health endpoint)
flask==3.0.0

# Data handling and utilities
python-dateutil==2.8.2

# HTTP client for health check testing (optional)
requests==2.31.0

